+++ functions/lib/shared/logging.js
@@ -1,0 +1,4 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.log = log;
+function log(event, fields) { console.info(JSON.stringify({ event, ...fields })); }